producer/requirements.txt
boto3==1.40.29
kafka-python==2.2.15
yfinance==0.2.65
pandas==2.3.0
numpy==2.3.1
python-dateutil==2.9.0.post0